+++ src/main/java/com/houssam/smartShop/dto/responseDTO/OrderResponseDTO.java
@@ -17,12 +17,12 @@ public class OrderResponseDTO {
 
     private String id;
     private LocalDate dateCommande;
-    private BigDecimal sousTotal;
-    private BigDecimal montantRemise;
-    private BigDecimal montantHT;
-    private BigDecimal tauxTVA;
-    private BigDecimal montantTVA;
-    private BigDecimal totalTTC;
+    private Double sousTotal;
+    private Double montantRemise;
+    private Double montantHT;
+    private Double tauxTVA;
+    private Double montantTVA;
+    private Double totalTTC;
     private OrderStatus statut;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
